web_app/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
langchain==0.1.10
langchain-openai==0.0.8
langchain-core==0.1.30
langgraph==0.0.20
openai>=1.40.0,<2.0.0
plotly==5.18.0
numpy>=1.26.4
python-dotenv==1.0.0
requests==2.31.0
reportlab==4.0.7
python-docx==1.1.0
pandas==2.3.3

yfinance
anthropic>=0.29.0
